feat: home화면 에러바운더리 처리 : react-error-boundary 패키지 추가 : Hero, 인기작 캐러셀UI  컴포넌트 분리 : 에러 발생시 대체 UI제공 위한 에러 바운더리 처리
