+++ package.json
@@ -28,6 +28,7 @@
     "lucide-react": "^0.477.0",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
+    "react-error-boundary": "^5.0.0",
     "react-hook-form": "^7.54.2",
     "react-router-dom": "^6.3.0",
     "supabase": "2.19.7",
