+++ src/components/PosterImage.tsx
@@ -7,7 +7,7 @@ const PosterImage = ({ posterPath, size }: posterImageType) => (
     {posterPath ? (
       <img src={getImageUrl(posterPath, size)} alt="movie_poster" className="w-full h-full" />
     ) : (
-      <div className="w-full  h-full bg-gray-200"> no poster </div>
+      <div className="absolute inset-0 w-full h-full bg-gray-200"> no poster </div>
     )}
   </>
 );
